requirements.txt
tensorflow-gpu
opencv-python
flask
numpy
Pillow
protobuf==3.20.*
streamlit
fastapi
uvicorn